requirements.txt
# Core dependencies
mcp>=0.9.0
pyserial>=3.5
protobuf>=4.25.0
pyyaml>=6.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
